module 28/requirements.txt
pytest
pytest-selenium
termcolor
allure-python-commons
selenium
python-dotenv
urllib3